% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/australia.post.codes.R
\docType{data}
\name{australia.post.codes}
\alias{australia.post.codes}
\title{Postcodes for Australia.}
\format{
A data.frame with 16,873 rows and 8 columns.
}
\source{
\url{https://download.geonames.org/export/zip/} and \url{https://wikipedia.org}
}
\usage{
data(euro.post.codes)
}
\description{
Postcodes, approximate latitude/longitude, state, suburb, region, and LGA info for cities and towns in Australia.
}
\details{
\itemize{
\item place - character string place name
\item post.code - character post code for \code{place}
\item state - Factor with 8 levels; state/1st order administrative subdivision
\item suburb - Character; suburb/locality/community geographic subdivision; some entries are missing.
\item latitude - Numeric, approximate latitude for region/place
\item longitude - Numeric, approximate longitude for region/place
\item lga - Local government area; third level of government in Australia; e.g. shires,
cities, town, municipal council, etc.
\item region - Factor with 79 levels containing the geographic region within each state for
\code{place}. N/A for external territories of Australia .
}
}
\keyword{datasets}
