% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reversegeocode.R
\name{ReverseGeocode}
\alias{ReverseGeocode}
\title{Reverse geocode latitude and longitude coordinates}
\usage{
ReverseGeocode(
  latitude,
  longitude,
  output.type = c("Admin1", "Country"),
  tol = set_units(100, "km")
)
}
\arguments{
\item{latitude}{Numeric vector of latitude coordinates (should be
between -90 and 90).}

\item{longitude}{Numeric vector of longitude coordinates (should be
between -180 and 180).}

\item{output.type}{String; either \code{"Country"} to return
country names or \code{"Admin1"} (the default) to return
state/province names.}

\item{tol}{Number giving the maximum distance in kilometers a point
may be from a region/polygon and still be declared a
match. Defaults to 100km, set to 0 to require all matches to be
inside the nearest polygon.}
}
\value{
A character vector containing the country or adminstrative
    region for each latitude-longitude pair. \code{"Other"} is
    returned for points outside where no matching country or region
    is found.
}
\description{
Converts input vectors of latitude and longitude coordinates to
their corresponding country or administrative region level 1
(state, province, etc.) within the country.
}
\examples{
ReverseGeocode(c(43.649, 37.768), c(-72.319, -75.666))
ReverseGeocode(c(47.906, 59.932), c(15.239, 30.197), output.type = "Country")
## Lighthouse just off the BC coast
ReverseGeocode(48.3917338,-123.3702094)
ReverseGeocode(48.3917338,-123.3702094, tol = 0)
}
