% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonyms.R
\docType{data}
\name{synonyms}
\alias{synonyms}
\title{Alternative names and ASCII names for several world regions, towns, and cities.}
\format{
A data.frame with 356,268 rows and 7 columns.
}
\source{
\url{https://download.geonames.org/export/dump/}
}
\usage{
data(synonyms)
}
\description{
Alternative names and ASCII names for several world regions, towns, and cities.
}
\details{
\itemize{
\item name - character string place name
\item ascii.name - ASCII character representation of \code{name}
\item alternate.names - Comma-separated string listing alternative names for the region/place
\item country.code - Factor with 18 levels; two-character country code for the country in ISO-3166 format.
\item latitude - Numeric, approximate latitude for region/place
\item longitude - Numeric, approximate longitude for region/place
\item time.zone - Factor with 84 levels containing the time zone for the region/place.
}
}
\keyword{datasets}
